edit

+++ controllers/BrandController.js
@@ -1,4 +1,6 @@
 const Brand = require("../models/BrandModel");
+const Product = require("../models/ProductModel");
+
 const slugify = require("slugify");
 
 const addBrand = async (req, res) => {
@@ -40,8 +42,16 @@ const getBrand = async (req, res) => {
 
 const deleteBrand = async (req, res) => {
   try {
-    const brand = await Brand.findByIdAndDelete(req.params.id);
-    res.status(200).json("Delete Successfully");
+    const brands = await Brand.findById(req.params.id);
+    const products = await Product.find({
+      nameBrand: brands.nameBrand,
+    });
+    if (products != "") {
+      res.status(211).json("Delete Fail");
+    } else {
+      const brand = await Brand.findByIdAndDelete(req.params.id);
+      res.status(200).json("Delete Successfully");
+    }
   } catch (error) {
     res.status(500).json(error);
   }
